Removed main thread query access to app database

+++ app/src/main/java/com/binay/recipeapp/data/repository/remote/RemoteRepo.kt
@@ -24,31 +24,36 @@ class RemoteRepo(private val apiHelper: ApiHelper, private val mDatabase: AppDat
 
     suspend fun getRecipes(tag: String): RecipeResponseData {
 
-        var recipeData = if (tag == "all")
+        val recipeData = if (tag == "all")
             apiHelper.getData("")
         else apiHelper.getData(tag)
 
 //        Fetch and remove previous recipes which are not favorite
         val recipeDao = mDatabase.recipeDao()
         val previousRecipes = recipeDao.getRecipes(tag)
-        if (previousRecipes != null) {
+
+        val isFirstLoad = previousRecipes?.isEmpty() == true
+        if (previousRecipes != null && !isFirstLoad) {
             val recipesWithNoFavoriteAndSameTag =
                 previousRecipes.filter { recipe -> recipe.tagToBeSearchedBy == tag && !recipe.isFavorite && !recipe.isRandom }
             val recipeIds = ArrayList<Int>()
             recipesWithNoFavoriteAndSameTag.forEach { recipe -> recipeIds.add(recipe.id) }
-            Log.e("Remove aghi Db Count ", " ${recipeDao.getRecipes(tag)?.count()}")
+            Log.e("Before Remove, Db Count ", " ${recipeDao.getRecipes(tag)?.count()}")
             recipeDao.removePreviousRecipes(recipeIds)
 
-            Log.e("Remove pachi Db Count ", " ${recipeDao.getRecipes(tag)?.count()}")
+            Log.e(" After Remove, Db Count ", " ${recipeDao.getRecipes(tag)?.count()}")
         }
+
         recipeData.recipes.forEach {
             it.tagToBeSearchedBy = tag
 //            Checks favorite Dao and updates data accordingly
 //                Note: If room has recipe, then it is automatically favorite
-            val favoriteRecipe = mDatabase.favoriteDao().getRecipe(it.id)
-            Log.e("Favorite Recipe: ", " $favoriteRecipe")
-            if (favoriteRecipe != null) {
-                it.isFavorite = true
+            if (!isFirstLoad) {
+                val favoriteRecipe = mDatabase.favoriteDao().getRecipe(it.id)
+                Log.e("Favorite Recipe: ", " $favoriteRecipe")
+                if (favoriteRecipe != null) {
+                    it.isFavorite = true
+                }
             }
         }
 //        Add new recipes
